pyproject.toml
[tool.poetry]
name = "horizon"
version = "0.1.0"
description = ""
authors = ["mrudisel <mitchrudisel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
matplotlib = "^3.8.4"
pillow = "^10.3.0"
pydantic = "^2.6.4"
tqdm = "^4.66.2"
tensorflow = "~2.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
